Add comment for bme.begin()
Clarification regarding the use of bme.begin() within the if statement. For those with more experience, this probably needs no explanation, but for someone with less experience such as myself this may save a bit of head scratching if someone decides their wiring is fine and don't need the if statement (because they will still need bme.begin() either way!).